src/backend/requirements.txt
flask
mysql-connector-python
python-dotenv
flask-cors
flask-bcrypt
flask-jwt-extended
gunicorn